[Chore] Service, Model, Helper로 각각 파일 분리 및 네이밍 수정
